requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
langchain-text-splitters
chromadb
pypdf
beautifulsoup4
requests
streamlit
python-dotenv
tavily-python
rank_bm25